pyproject.toml
[tool.poetry]
name = "gifer"
version = "0.1.0"
description = "Screenshot to GIF/Video Service"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = {extras = ["standard"], version = "^0.35.0"}
playwright = "^1.55.0"
Pillow = "^11.3.0"
imageio = "^2.37.0"
structlog = "^25.4.0"
tenacity = "^9.1.2"
aiolimiter = "^1.2.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
typing-extensions = "^4.15.0"
google-api-python-client = "^2.179.0"
google-auth = "^2.40.3"
python-dotenv = "^1.1.1"

[tool.poetry.dev-dependencies]
ruff = "^0.12.11"
black = "^25.1.0"
mypy = "^1.17.1"
pytest = "^8.4.1"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
